requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pdfminer.six==20221105
python-docx==0.8.11
openpyxl==3.1.2
pandas==2.0.3
PyPDF2==3.0.1
tldextract==3.4.4
sqlalchemy==2.0.19
python-slugify==8.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.1.0
poppler-utils  # Note: This needs to be installed separately on system
opencv-python-headless==4.9.0.80
numpy==1.26.4
spacy 
transformers 
torch 
py2neo 
tqdm
neo4j==5.12.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==0.28.0
python-dotenv==1.0.0
numpy==1.24.3
tqdm==4.66.1
#python -m spacy download en_core_web_lg
